requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cpu

# torch==2.9.1+cpu
# sentence-transformers==2.7.0

langchain
langchain-openai
# langchain-text-splitters
# langgraph
pydantic
pandas

# chromadb
python-dotenv
# pypdf
fastapi
uvicorn

python-multipart

# unstructured[pdf]

pytest
pytest-asyncio
httpx

google-cloud-bigquery

openai